back/requirements.txt
Flask==2.3.3
pymongo==4.5.0
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
certifi
flask-socketio==5.3.6
schedule==1.2.0
twilio>=8.10.0
requests>=2.31.0